+++ system/lib/compiler-rt/lib/sanitizer_common/sanitizer_posix_libcdep.cc
@@ -305,7 +305,9 @@ void PlatformPrepareForSandboxing(__sanitizer_sandbox_arguments *args) {
   // to read the file mappings from /proc/self/maps. Luckily, neither the
   // process will be able to load additional libraries, so it's fine to use the
   // cached mappings.
+#if !SANITIZER_EMSCRIPTEN
   MemoryMappingLayout::CacheMemoryMappings();
+#endif
 }
 
 #if SANITIZER_ANDROID || SANITIZER_GO