+++ tools/shared.py
@@ -22,12 +22,12 @@ import subprocess
 import sys
 import tempfile
 
-if sys.version_info < (2, 7, 12):
-  print('emscripten requires python 2.7.12 or above', file=sys.stderr)
+if sys.version_info < (2, 7, 0):
+  print('emscripten requires python 2.7.0 or above (python 2.7.12 or newer is recommended, older python versions are known to run into SSL related issues, https://github.com/emscripten-core/emscripten/issues/6275)', file=sys.stderr)
   sys.exit(1)
 
 if sys.version_info[0] == 3 and sys.version_info < (3, 5):
-  print('emscripten requires at least python 3.5 (or python 2.7.12)', file=sys.stderr)
+  print('emscripten requires at least python 3.5 (or python 2.7.12 or above)', file=sys.stderr)
   sys.exit(1)
 
 from .toolchain_profiler import ToolchainProfiler
@@ -46,6 +46,9 @@ logging.basicConfig(format='%(name)s:%(levelname)s: %(message)s',
                     level=logging.DEBUG if DEBUG else logging.INFO)
 colored_logger.enable()
 logger = logging.getLogger('shared')
+
+if sys.version_info < (2, 7, 12):
+  logger.debug('python versions older than 2.7.12 are known to run into outdated SSL certificate related issues, https://github.com/emscripten-core/emscripten/issues/6275')
 
 
 def exit_with_error(msg, *args):